Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/mediacapture-streams/MediaStream-feature-policy-none.https-expected.txt

CONSOLE MESSAGE: Feature policy 'Camera' check failed for iframe with origin 'https://localhost:9443' and allow attribute 'camera 'none''.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Camera' check failed for iframe with origin 'https://localhost:9443' and allow attribute 'camera 'none''.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Microphone' check failed for iframe with origin 'https://127.0.0.1:9443' and allow attribute 'camera 'none''.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Camera' check failed for iframe with origin 'https://127.0.0.1:9443' and allow attribute 'camera 'none''.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Microphone' check failed for iframe with origin 'https://127.0.0.1:9443' and allow attribute 'camera 'none''.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Microphone' check failed for iframe with origin 'https://localhost:9443' and allow attribute 'microphone 'none''.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Microphone' check failed for iframe with origin 'https://localhost:9443' and allow attribute 'microphone 'none''.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Microphone' check failed for iframe with origin 'https://localhost:9443' and allow attribute 'microphone 'none';camera 'none''.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Camera' check failed for iframe with origin 'https://localhost:9443' and allow attribute 'microphone 'none';camera 'none''.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Microphone' check failed for iframe with origin 'https://localhost:9443' and allow attribute 'microphone 'none';camera 'none''.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Microphone' check failed for iframe with origin 'https://localhost:9443' and allow attribute 'camera *;microphone 'none''.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Microphone' check failed for iframe with origin 'https://localhost:9443' and allow attribute 'camera *;microphone 'none''.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Microphone' check failed for iframe with origin 'https://127.0.0.1:9443' and allow attribute 'camera *;microphone 'none''.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Microphone' check failed for iframe with origin 'https://127.0.0.1:9443' and allow attribute 'camera *;microphone 'none''.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Camera' check failed for iframe with origin 'https://localhost:9443' and allow attribute 'camera 'none';microphone *'.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Camera' check failed for iframe with origin 'https://localhost:9443' and allow attribute 'camera 'none';microphone *'.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Microphone' check failed for iframe with origin 'https://localhost:9443' and allow attribute 'microphone *; microphone 'none''.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Microphone' check failed for iframe with origin 'https://localhost:9443' and allow attribute 'microphone *; microphone 'none''.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Camera' check failed for iframe with origin 'https://127.0.0.1:9443' and allow attribute 'microphone *; camera 'self''.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Camera' check failed for iframe with origin 'https://127.0.0.1:9443' and allow attribute 'microphone *; camera 'self''.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Camera' check failed for iframe with origin 'https://localhost:9443' and allow attribute 'microphone *; camera http:/example.org self'.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Camera' check failed for iframe with origin 'https://localhost:9443' and allow attribute 'microphone *; camera http:/example.org self'.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Camera' check failed for iframe with origin 'https://127.0.0.1:9443' and allow attribute 'microphone *; camera http:/example.org 'self''.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Camera' check failed for iframe with origin 'https://127.0.0.1:9443' and allow attribute 'microphone *; camera http:/example.org 'self''.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Microphone' check failed for iframe with origin 'https://localhost:9443' and allow attribute 'microphone https://127.0.0.1:9443'.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Microphone' check failed for iframe with origin 'https://localhost:9443' and allow attribute 'microphone https://127.0.0.1:9443'.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Camera' check failed for iframe with origin 'https://127.0.0.1:9443' and allow attribute 'microphone https://127.0.0.1:9443'.
CONSOLE MESSAGE: Not allowed to call getUserMedia.
CONSOLE MESSAGE: Feature policy 'Camera' check failed for iframe with origin 'https://127.0.0.1:9443' and allow attribute 'microphone https://127.0.0.1:9443'.
CONSOLE MESSAGE: Not allowed to call getUserMedia.

PASS "camera 'none'" - same origin iframe
PASS "camera 'none'" - cross origin iframe
PASS "microphone 'none'" - same origin iframe
PASS "microphone 'none';camera 'none'" - same origin iframe
PASS "camera *;microphone 'none'" - same origin iframe
PASS "camera *;microphone 'none'" - cross origin iframe
PASS "camera 'none';microphone *" - same origin iframe
PASS "microphone *; microphone 'none'" - same origin iframe
PASS "microphone *; microphone ' none'" - same origin iframe
PASS "microphone *; microphone none" - same origin iframe
PASS "microphone *; camera 'self'" - same origin iframe
PASS "microphone *; camera 'self'" - cross origin iframe
PASS "microphone *; camera http:/example.org self" - same origin iframe
PASS "microphone *; camera http:/example.org 'self'" - same origin iframe
PASS "microphone *; camera http:/example.org 'self'" - cross origin iframe
PASS "microphone https://127.0.0.1:9443" - same origin iframe
PASS "microphone https://127.0.0.1:9443" - cross origin iframe
PASS "microphone 'self' https://127.0.0.1:9443" - same origin iframe
